Refactored the transaction_payment pallets
