+++ runtime/src/configs/mod.rs
@@ -1,2 +1,3 @@
 pub mod timestamp;
 pub mod balances;
+pub mod transaction_payment;
